chore: move ProposalEnum under its own module.
